backend/requirements.txt
openai
groq
flask
requests
python-dotenv
aiohttp
flask-cors
flask[async]
requests
firebase-admin
google-cloud-storage